Modal excluir funcionando

+++ pages/config/exclui_atividade.php
@@ -4,18 +4,18 @@
 	<title>Exlui Atividade</title>
 	</head>
 	<body>
-		<?php 
+	<?php 
 	$conn = mysql_connect("localhost","root","") or die("ERRO FUUUUH!");
 	mysql_select_db("dbcci");
 	$exc = $_POST['excluir'];
-	//if(isset($exc)){
+	$tipo = $_GET['tipo_atividade'];
+	if(isset($exc)){
 		$id = $_GET['id'];
 		echo $id;
 		mysql_query("DELETE FROM tb_atividade WHERE id_atividade='".$id."'");
 		mysql_close($conn);
-		
-		header("location:lista_atividade.php");
-//	}
+	}
+		header("location:lista_envento.php");
 ?>
 	</body>
